move calculate_metrics.py
