+++ supervisely/nn/benchmark/coco_utils/__init__.py
@@ -1,2 +1,3 @@
 from supervisely.nn.benchmark.coco_utils.sly2coco import sly2coco
 from supervisely.nn.benchmark.coco_utils.utils import read_coco_datasets
+from supervisely.nn.benchmark.coco_utils.calculate_metrics import calculate_metrics
