--- supervisely/nn/benchmark/evaluation/instance_segmentation/__init__.py
@@ -1,2 +1,0 @@
-from supervisely.nn.benchmark.evaluation.instance_segmentation.calculate_metrics import calculate_metrics
-# from supervisely.nn.benchmark.evaluation.object_detection.metric_provider import
