+++ supervisely/nn/benchmark/evaluation/object_detection/__init__.py
@@ -1,2 +1,1 @@
-from supervisely.nn.benchmark.evaluation.object_detection.calculate_metrics import calculate_metrics
 from supervisely.nn.benchmark.evaluation.object_detection.metric_provider import MetricProvider
